refactor(buildsource): extract merge_support; fix A2 digest order (CI+Codex)
CI: A2 pushed cli_buildsource.py over the 2000-line hard cap (2088),
failing the file-size ai-readiness gate (and test_ai_readiness on every
platform). Extract the pack-combination + merge-conflict helpers
(_combine_packs, _filter_pack_layers, _layer_value, _detect/_record merge
conflicts) into abicheck/buildsource/merge_support.py; cli_buildsource is
back to 1889 lines.

Codex P2: the A2 conflict digest hashed to_dict() in serializer order, so
two inputs with identical facts in a different list order (e.g. reversed
compile_commands.json) falsely conflicted and --on-conflict=error would
abort a valid baseline. Add _canonicalize (recursively sort dict keys +
lists by canonical JSON) and _canonical_layer_digest; the conflict
detector now uses an order-independent digest. _combine_packs's
content-hash path is left untouched.

Test: reversed compile DB is not a conflict under --on-conflict=error.
