+++ abicheck/cli_buildsource.py
@@ -25,7 +25,6 @@ explicit opt-in not implemented here.
 from __future__ import annotations
 
 import datetime as _dt
-from dataclasses import replace
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -40,6 +39,13 @@ from .buildsource.evidence_policy import (
     finding_bucket_counts,
     require_evidence_findings,
     tag_evidence_category,
+)
+from .buildsource.merge_support import (
+    _combine_packs,
+    _detect_merge_layer_conflicts,
+    _filter_pack_layers,
+    _layer_value,
+    _record_merge_conflicts,
 )
 from .buildsource.model import (
     CoverageStatus,
@@ -1048,211 +1054,6 @@ def _collect_source_abi_android(
     )
 
 # ── Attach / compare integration (ADR-028 D6, D7; ADR-029 D9) ─────────────────
-
-def _layer_value(layer: object) -> str:
-    return layer.value if hasattr(layer, "value") else str(layer)
-
-def _filter_pack_layers(
-    pack: BuildSourcePack | None, layers: tuple[str, ...]
-) -> BuildSourcePack | None:
-    """Null out a loaded pack's facts for layers the collect-mode excludes, so a
-    pre-captured pack can't smuggle past the ADR-033 D2 layer set (Codex review).
-    ``_combine_packs`` derives coverage from these attributes, so nulling them
-    drops both the facts and their coverage rows."""
-    if pack is None:
-        return None
-    if "L3" not in layers:
-        pack.build_evidence = None
-    if "L4" not in layers:
-        pack.source_abi = None
-    if "L5" not in layers:
-        pack.source_graph = None
-    return pack
-
-def _combine_packs(
-    bi_pack: BuildSourcePack | None,
-    src_pack: BuildSourcePack | None,
-    embedded: BuildSourcePack | None = None,
-) -> BuildSourcePack | None:
-    """Combine a build-info pack and a sources pack into one embeddable pack.
-
-    Facts are taken from the pack that supplies each layer — ``build_evidence``
-    from ``--build-info``, ``source_abi``/``source_graph`` from ``--sources`` —
-    with *embedded* backfilling any gap. The coverage manifest is rebuilt by
-    pulling each layer's row from the *same* pack that supplied its facts (not
-    just the base pack), then dropping rows for layers we do not actually carry.
-    This keeps a later compare's coverage/capability report honest when the two
-    flags point at different packs (Codex review). Returns ``None`` when no pack
-    contributes any facts.
-    """
-    def _first(attr: str, *packs: BuildSourcePack | None) -> object | None:
-        for cand in packs:
-            if cand is not None and getattr(cand, attr) is not None:
-                return getattr(cand, attr)
-        return None
-
-    build_evidence = _first("build_evidence", bi_pack, src_pack, embedded)
-    source_abi = _first("source_abi", src_pack, bi_pack, embedded)
-    source_graph = _first("source_graph", src_pack, bi_pack, embedded)
-
-    base = bi_pack or src_pack or embedded
-    if base is None:
-        return None
-
-    # supplier order per managed layer, and whether we actually carry it
-    supplier: dict[str, tuple[BuildSourcePack | None, ...]] = {
-        DataLayer.L3_BUILD.value: (bi_pack, src_pack, embedded),
-        DataLayer.L4_SOURCE_ABI.value: (src_pack, bi_pack, embedded),
-        DataLayer.L5_SOURCE_GRAPH.value: (src_pack, bi_pack, embedded),
-    }
-    present = {
-        DataLayer.L3_BUILD.value: build_evidence is not None,
-        DataLayer.L4_SOURCE_ABI.value: source_abi is not None,
-        DataLayer.L5_SOURCE_GRAPH.value: source_graph is not None,
-    }
-    managed = set(supplier)
-
-    coverage: list[LayerCoverage] = []
-    # Non-managed rows (L0/L1/L2/…) come from the base manifest.
-    for c in base.manifest.coverage:
-        if _layer_value(c.layer) not in managed:
-            coverage.append(c)
-    # Always emit one row per managed layer (ADR-028 D7 shows every layer). When
-    # we carry the facts, reuse the supplying pack's row; otherwise mark the
-    # layer not_collected so the report never advertises a check with no facts
-    # behind it (Codex review) — and never drops the row entirely either.
-    for layer, packs in supplier.items():
-        row: LayerCoverage | None = None
-        if present[layer]:
-            for cand in packs:
-                if cand is None:
-                    continue
-                row = next(
-                    (c for c in cand.manifest.coverage if _layer_value(c.layer) == layer),
-                    None,
-                )
-                if row is not None:
-                    break
-            if row is None:
-                row = LayerCoverage(layer=layer, status=CoverageStatus.PRESENT)
-        else:
-            row = LayerCoverage(layer=layer, status=CoverageStatus.NOT_COLLECTED)
-        coverage.append(row)
-
-    # Provenance: the combined manifest's artifacts/extractors must reflect every
-    # pack that supplied an embedded fact, not just the base pack — otherwise
-    # to_ref()/content_hash() would omit the source pack's artifacts for a
-    # cross-pack self-contained snapshot (Codex review). A pack "contributed"
-    # when one of its facts is the object we actually embedded.
-    chosen = (build_evidence, source_abi, source_graph)
-    chosen_ids = {id(x) for x in chosen if x is not None}
-    artifacts: list[str] = []
-    extractors: list[ExtractorRecord] = []
-    seen_extractors: set[tuple[str, str]] = set()
-    for p in (bi_pack, src_pack, embedded):
-        if p is None or not (
-            chosen_ids & {id(p.build_evidence), id(p.source_abi), id(p.source_graph)}
-        ):
-            continue
-        for a in p.manifest.artifacts:
-            if a not in artifacts:
-                artifacts.append(a)
-        for e in p.manifest.extractors:
-            key = (e.name, e.version)
-            if key not in seen_extractors:
-                seen_extractors.add(key)
-                extractors.append(e)
-    # An *inline*-collected contributor (e.g. `--sources <raw tree>`) is never
-    # written to disk, so its manifest.artifacts is empty and the loop above
-    # adds no digest for its source_abi/source_graph. Since content_hash() trusts
-    # a non-empty manifest.artifacts, a mixed `--build-info <pack> --sources
-    # <tree>` would then hash only the build pack's digest and ignore the source
-    # facts — two different trees with the same build pack collide (Codex P2). Add
-    # the in-memory payload digest for every chosen fact; a fact that *was*
-    # written to disk hashes identically (_payload_sha256 mirrors _write_json), so
-    # it dedups against the on-disk digest above rather than double-counting.
-    from .buildsource.pack import _payload_sha256
-
-    for payload in chosen:
-        if payload is None:
-            continue
-        digest = "sha256:" + _payload_sha256(payload.to_dict())  # type: ignore[attr-defined]
-        if digest not in artifacts:
-            artifacts.append(digest)
-
-    return BuildSourcePack(
-        root=Path(""),
-        manifest=replace(
-            base.manifest, coverage=coverage, artifacts=artifacts, extractors=extractors
-        ),
-        build_evidence=build_evidence,  # type: ignore[arg-type]
-        source_abi=source_abi,  # type: ignore[arg-type]
-        source_graph=source_graph,  # type: ignore[arg-type]
-    )
-
-_MERGE_LAYER_ATTRS: dict[str, str] = {
-    DataLayer.L3_BUILD.value: "build_evidence",
-    DataLayer.L4_SOURCE_ABI.value: "source_abi",
-    DataLayer.L5_SOURCE_GRAPH.value: "source_graph",
-}
-
-def _detect_merge_layer_conflicts(
-    snaps: list[tuple[Path, AbiSnapshot]],
-) -> dict[str, list[tuple[str, str]]]:
-    """A2: per managed layer, return ``layer -> [(input_name, digest), ...]`` when
-    >1 input supplies that layer with *differing* normalized facts.
-
-    The comparison is a **per-layer payload digest** of just that layer's facts,
-    not the pack-wide ``BuildSourcePack.content_hash()`` — the pack hash folds in
-    every layer plus coverage/extractor metadata, so two inputs with identical
-    L4/L5 facts but a differing unrelated layer would false-positive. A layer with
-    one contributor, or several contributors that all agree, is not a conflict.
-    """
-    from .buildsource.pack import _payload_sha256
-
-    seen: dict[str, list[tuple[str, str]]] = {layer: [] for layer in _MERGE_LAYER_ATTRS}
-    for path, s in snaps:
-        pack = s.build_source
-        if pack is None:
-            continue
-        for layer, attr in _MERGE_LAYER_ATTRS.items():
-            payload = getattr(pack, attr, None)
-            if payload is None:
-                continue
-            digest = "sha256:" + _payload_sha256(payload.to_dict())
-            seen[layer].append((path.name, digest))
-
-    conflicts: dict[str, list[tuple[str, str]]] = {}
-    for layer, entries in seen.items():
-        if len(entries) > 1 and len({d for _n, d in entries}) > 1:
-            conflicts[layer] = entries
-    return conflicts
-
-def _record_merge_conflicts(
-    combined: BuildSourcePack, conflicts: dict[str, list[tuple[str, str]]]
-) -> None:
-    """Persist A2 conflicts into the combined pack's extractor ledger.
-
-    ``BuildSourceManifest.to_dict()`` serializes ``extractors`` (but has no
-    ``diagnostics`` field), so an ``ExtractorRecord`` is the channel that
-    survives embedding/round-trip. ``warn`` mode keeps first-wins facts and
-    leaves this record behind so the divergence rides forward in the baseline.
-    """
-    records = list(combined.manifest.extractors)
-    for layer, entries in sorted(conflicts.items()):
-        detail = "; ".join(f"{name}={digest}" for name, digest in entries)
-        records.append(
-            ExtractorRecord(
-                name="merge_layer_conflict",
-                status="failed",
-                detail=f"layer {layer} supplied with differing facts: {detail}",
-                diagnostics=[
-                    f"kept first-wins for {layer}; verify each layer comes from "
-                    "exactly one input."
-                ],
-            )
-        )
-    combined.manifest = replace(combined.manifest, extractors=records)
 
 def embed_build_source(
     snap: AbiSnapshot,
